Improve single responsibility of function
It was rather counter-intuitive and error-prone that function
responsible of posting comment to last movie actually posted the movie,
too.

It is now _not_ protected against scenario when there is no movie in the
database during a call, but such "design by contract" seems better than
previous solution, because in such scenario there will be the
AttributeError, which should be easy to detect, especially in TestCase.